requirements.txt
2captcha-python==1.2.2
attrs==23.2.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
exceptiongroup==1.2.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gspread==6.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
pyparsing==3.1.1
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
StrEnum==0.4.15
trio==0.24.0
trio-websocket==0.11.1
urllib3==1.26.18
urllib3-secure-extra==0.1.0
webdriver-manager==4.0.1
wsproto==1.2.0
